refactor: protected route refactor
