+++ frontend/src/components/Unauthorized.jsx
@@ -1,0 +1,8 @@
+const Unauthorized = () => (
+    <div>
+        <h1>Unauthorized</h1>
+        <p>You do not have permission to view this page.</p>
+    </div>
+);
+
+export default Unauthorized;
